remove polyfill-3d (#5425)
