+++ cocos2d/core/index.js
@@ -44,6 +44,5 @@ if (!CC_EDITOR || !Editor.isMainProcess) {
 
 require('./mesh');
 require('./3d');
-require('./3d/polyfill-3d');
 
 require('./base-ui/CCWidgetManager');
